added card view of records
